- use "int" instead of "ssize_t" for buffer size

+++ HomerMultimedia/include/MediaSinkMem.h
@@ -52,7 +52,7 @@ public:
 
     virtual ~MediaSinkMem();
 
-    virtual void ReadFragment(char *pData, ssize_t &pDataSize);
+    virtual void ReadFragment(char *pData, int &pDataSize);
     virtual void StopReading();
 
 protected:
